requirements.txt
# Web framework
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Telegram bot
python-telegram-bot[ext]==21.6

# HTTP client
httpx==0.27.2

# Database (async)
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
greenlet==3.1.1

# Scheduling
APScheduler==3.10.4

# Environment + config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# AI providers
anthropic==0.39.0
openai==1.54.3

# Email
resend==2.4.0

# Observability dashboard push
PyGithub==2.4.0

# Timezones
pytz==2024.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
